refactor(gateway): 重构网关模块结构，统一broker实现与组件拆分
将各券商网关实现迁移至brokers子目录，拆分ctp、miniqmt、ptrade为独立子包；拆分网关核心协议到protocols.py，原base.py改为兼容导出层；新增事件桥接、故障恢复等独立功能模块，重命名models为broker_models并保留兼容导入；更新所有相关导入路径、文档与测试用例，保持对外接口兼容。
