+++ python/akquant/gateway/brokers/__init__.py
@@ -1,0 +1,1 @@
+"""Broker-specific gateway implementations."""
