+++ python/akquant/gateway/brokers/miniqmt/__init__.py
@@ -1,0 +1,3 @@
+from .stub import MiniQMTMarketGateway, MiniQMTTraderGateway
+
+__all__ = ["MiniQMTMarketGateway", "MiniQMTTraderGateway"]
