+++ python/akquant/gateway/order_submitter.py
@@ -1,0 +1,268 @@
+from typing import Any, Callable
+
+from ..log import build_log_extra, get_logger
+from .broker_models import (
+    BrokerCapability,
+    UnifiedOrderRequest,
+    validate_execution_semantics,
+)
+
+logger = get_logger("gateway.live")
+
+
+def find_live_close_position(
+    positions: Any,
+    symbol: str,
+    side: str,
+    payload_field: Callable[[Any, str], Any],
+) -> Any | None:
+    """Find the position that should be reduced by the given order side."""
+    target_symbol = str(symbol).strip()
+    normalized_side = str(side).strip().lower()
+    for position in positions or ():
+        position_symbol = str(payload_field(position, "symbol")).strip()
+        if position_symbol != target_symbol:
+            continue
+        position_direction = str(payload_field(position, "direction")).strip().lower()
+        position_quantity = float(payload_field(position, "quantity") or 0.0)
+        if normalized_side == "sell" and (
+            position_direction in {"buy", "long"} or position_quantity > 0
+        ):
+            return position
+        if normalized_side == "buy" and (
+            position_direction in {"sell", "short"} or position_quantity < 0
+        ):
+            return position
+    return None
+
+
+def resolve_live_order_legs(
+    trader_gateway: Any,
+    capability: BrokerCapability,
+    symbol: str,
+    side: str,
+    quantity: float,
+    position_effect: str,
+    reduce_only: bool,
+    payload_field: Callable[[Any, str], Any],
+) -> list[tuple[str, float]]:
+    """Resolve close orders into close_today/close_yesterday legs when supported."""
+    normalized_effect = str(position_effect).strip().lower()
+    if quantity <= 0 or reduce_only or normalized_effect != "close":
+        return [(normalized_effect, quantity)]
+    supported_effects = {
+        str(item).strip().lower() for item in capability.supported_position_effects
+    }
+    if not capability.position_details:
+        return [(normalized_effect, quantity)]
+    if (
+        "close_today" not in supported_effects
+        or "close_yesterday" not in supported_effects
+    ):
+        return [(normalized_effect, quantity)]
+
+    query_positions = getattr(trader_gateway, "query_positions", None)
+    if not callable(query_positions):
+        return [(normalized_effect, quantity)]
+    try:
+        positions = query_positions()
+    except Exception:
+        return [(normalized_effect, quantity)]
+
+    target_position = find_live_close_position(positions, symbol, side, payload_field)
+    if target_position is None:
+        return [(normalized_effect, quantity)]
+
+    available_today = max(
+        0.0, float(getattr(target_position, "available_today_quantity", 0.0) or 0.0)
+    )
+    available_yesterday = max(
+        0.0,
+        float(getattr(target_position, "available_yesterday_quantity", 0.0) or 0.0),
+    )
+    split_quantity = min(quantity, available_today + available_yesterday)
+    legs: list[tuple[str, float]] = []
+    close_today_quantity = min(split_quantity, available_today)
+    if close_today_quantity > 0:
+        legs.append(("close_today", close_today_quantity))
+    remaining_quantity = max(split_quantity - close_today_quantity, 0.0)
+    close_yesterday_quantity = min(remaining_quantity, available_yesterday)
+    if close_yesterday_quantity > 0:
+        legs.append(("close_yesterday", close_yesterday_quantity))
+    unresolved_quantity = max(quantity - split_quantity, 0.0)
+    if unresolved_quantity > 0 or not legs:
+        legs.append(
+            ("close", unresolved_quantity if unresolved_quantity > 0 else quantity)
+        )
+    return legs
+
+
+def build_live_order_client_ids(
+    request_client_order_id: str, order_legs: list[tuple[str, float]]
+) -> list[str]:
+    """Build per-leg client order ids while keeping the first leg stable."""
+    if len(order_legs) <= 1:
+        return [request_client_order_id]
+    client_order_ids = [request_client_order_id]
+    for leg_index, (position_effect, _) in enumerate(order_legs[1:], start=2):
+        suffix = str(position_effect).replace("_", "-")
+        client_order_ids.append(f"{request_client_order_id}-{suffix}-{leg_index}")
+    return client_order_ids
+
+
+def validate_live_order_client_ids(
+    strategy: Any,
+    client_order_ids: list[str],
+    symbol: str,
+    side: str,
+    quantity: float,
+    can_submit_client_order: Callable[[str], bool],
+    notify_strategy_error: Callable[[Any, Exception, str, Any], None],
+) -> None:
+    """Ensure all generated client order ids are available before submit."""
+    owner_strategy_id = str(getattr(strategy, "_owner_strategy_id", "_default")).strip()
+    owner_strategy_id = owner_strategy_id or "_default"
+    for client_order_id in client_order_ids:
+        if can_submit_client_order(client_order_id):
+            continue
+        exc = RuntimeError(f"duplicate active client_order_id: {client_order_id}")
+        logger.warning(
+            "Rejected live submit_order because client_order_id is already active",
+            extra=build_log_extra(
+                phase="gateway",
+                strategy_id=owner_strategy_id,
+                slot=owner_strategy_id if owner_strategy_id != "_default" else None,
+                symbol=symbol,
+                client_order_id=client_order_id,
+            ),
+        )
+        notify_strategy_error(
+            strategy,
+            exc,
+            "submit_order",
+            {
+                "client_order_id": client_order_id,
+                "symbol": symbol,
+                "side": side,
+                "quantity": quantity,
+            },
+        )
+        raise exc
+
+
+class BrokerOrderSubmitter:
+    """Inject broker-live submit_order helpers into a strategy instance."""
+
+    def __init__(
+        self,
+        *,
+        trader_gateway: Any,
+        strategy: Any,
+        resolve_trader_capabilities: Callable[[Any], BrokerCapability],
+        next_client_order_id: Callable[[], str],
+        can_submit_client_order: Callable[[str], bool],
+        sync_order_id_mapping: Callable[[str, str], None],
+        bind_order_owner: Callable[[str, str, str], None],
+        notify_strategy_error: Callable[[Any, Exception, str, Any], None],
+        payload_field: Callable[[Any, str], Any],
+        get_execution_capabilities: Callable[[], dict[str, Any]],
+    ) -> None:
+        """Bind strategy injection hooks, id mapping and capability callbacks."""
+        self._trader_gateway = trader_gateway
+        self._strategy = strategy
+        self._resolve_trader_capabilities = resolve_trader_capabilities
+        self._next_client_order_id = next_client_order_id
+        self._can_submit_client_order = can_submit_client_order
+        self._sync_order_id_mapping = sync_order_id_mapping
+        self._bind_order_owner = bind_order_owner
+        self._notify_strategy_error = notify_strategy_error
+        self._payload_field = payload_field
+        self._get_execution_capabilities = get_execution_capabilities
+
+    def install(self) -> None:
+        """Bind broker-live submit_order helpers onto the strategy object."""
+        setattr(self._strategy, "submit_order", self.submit_order)
+        setattr(
+            self._strategy,
+            "can_submit_client_order",
+            self._can_submit_client_order,
+        )
+        setattr(
+            self._strategy,
+            "get_execution_capabilities",
+            self._get_execution_capabilities,
+        )
+
+    def submit_order(
+        self,
+        symbol: str,
+        side: str,
+        quantity: float,
+        price: float | None = None,
+        client_order_id: str | None = None,
+        order_type: str = "Market",
+        time_in_force: str = "GTC",
+        trigger_price: float | None = None,
+        tag: str | None = None,
+        position_effect: str = "auto",
+        reduce_only: bool = False,
+        extra: dict[str, Any] | None = None,
+    ) -> str:
+        """Submit a live broker order using the unified strategy-facing signature."""
+        _ = trigger_price
+        _ = tag
+        if extra:
+            raise RuntimeError("extra broker fields are not supported")
+        capability = self._resolve_trader_capabilities(self._trader_gateway)
+        normalized_position_effect = validate_execution_semantics(
+            capability,
+            position_effect,
+            reduce_only,
+        )
+        owner_strategy_id = str(
+            getattr(self._strategy, "_owner_strategy_id", "_default")
+        )
+        order_legs = resolve_live_order_legs(
+            trader_gateway=self._trader_gateway,
+            capability=capability,
+            symbol=symbol,
+            side=side,
+            quantity=quantity,
+            position_effect=normalized_position_effect,
+            reduce_only=reduce_only,
+            payload_field=self._payload_field,
+        )
+        request_client_order_id = client_order_id or self._next_client_order_id()
+        client_order_ids = build_live_order_client_ids(
+            request_client_order_id=request_client_order_id,
+            order_legs=order_legs,
+        )
+        validate_live_order_client_ids(
+            strategy=self._strategy,
+            client_order_ids=client_order_ids,
+            symbol=symbol,
+            side=side,
+            quantity=quantity,
+            can_submit_client_order=self._can_submit_client_order,
+            notify_strategy_error=self._notify_strategy_error,
+        )
+        broker_order_ids: list[str] = []
+        for leg_index, (leg_position_effect, leg_quantity) in enumerate(order_legs):
+            request = UnifiedOrderRequest(
+                client_order_id=client_order_ids[leg_index],
+                symbol=symbol,
+                side=side,
+                quantity=leg_quantity,
+                price=price,
+                order_type=order_type,
+                time_in_force=time_in_force,
+                position_effect=leg_position_effect,
+                reduce_only=reduce_only,
+            )
+            broker_order_id = str(self._trader_gateway.place_order(request))
+            broker_order_ids.append(broker_order_id)
+            self._sync_order_id_mapping(request.client_order_id, broker_order_id)
+            self._bind_order_owner(
+                request.client_order_id, broker_order_id, owner_strategy_id
+            )
+        return broker_order_ids[0]
