+++ tests/test_live_runner_broker_bridge.py
@@ -3,6 +3,7 @@ from pathlib import Path
 from types import SimpleNamespace
 from typing import Any, Callable, cast
 
+import akquant.live as live_module
 import pytest
 from akquant.gateway.models import BrokerCapability, UnifiedPosition
 from akquant.live import LiveRunner
@@ -405,6 +406,84 @@ def test_live_runner_compatible_recovery_keeps_sync_failure_silent() -> None:
 
     assert strategy.errors == []
     assert broker_events == []
+
+
+def test_live_runner_init_normalizes_legacy_gateway_options(
+    monkeypatch: pytest.MonkeyPatch,
+) -> None:
+    """LiveRunner should fold legacy broker args into gateway_options at init."""
+
+    class _DummyDataFeed:
+        @staticmethod
+        def create_live() -> object:
+            return object()
+
+    class _DummyEngine:
+        pass
+
+    monkeypatch.setattr(live_module, "DataFeed", _DummyDataFeed)
+    monkeypatch.setattr(live_module, "Engine", _DummyEngine)
+
+    runner = LiveRunner(
+        strategy_cls=None,
+        instruments=[],
+        md_front="tcp://md-front",
+        broker_id="9999",
+        user_id="trader-a",
+        password="secret",
+        gateway_options={"recovery_mode": "strict", "custom": "value"},
+    )
+
+    assert runner.gateway_options == {
+        "recovery_mode": "strict",
+        "custom": "value",
+        "md_front": "tcp://md-front",
+        "broker_id": "9999",
+        "user_id": "trader-a",
+        "password": "secret",
+    }
+    assert runner.md_front == "tcp://md-front"
+    assert runner.broker_id == "9999"
+    assert runner.user_id == "trader-a"
+    assert runner.password == "secret"
+    assert runner._build_gateway_kwargs() == runner.gateway_options
+
+
+def test_live_runner_init_prefers_gateway_options_over_legacy_args(
+    monkeypatch: pytest.MonkeyPatch,
+) -> None:
+    """Explicit gateway_options values should win over legacy init args."""
+
+    class _DummyDataFeed:
+        @staticmethod
+        def create_live() -> object:
+            return object()
+
+    class _DummyEngine:
+        pass
+
+    monkeypatch.setattr(live_module, "DataFeed", _DummyDataFeed)
+    monkeypatch.setattr(live_module, "Engine", _DummyEngine)
+
+    runner = LiveRunner(
+        strategy_cls=None,
+        instruments=[],
+        md_front="tcp://legacy-md-front",
+        broker_id="legacy-broker",
+        user_id="legacy-user",
+        gateway_options={
+            "md_front": "tcp://explicit-md-front",
+            "broker_id": "explicit-broker",
+            "user_id": "explicit-user",
+        },
+    )
+
+    assert runner.gateway_options["md_front"] == "tcp://explicit-md-front"
+    assert runner.gateway_options["broker_id"] == "explicit-broker"
+    assert runner.gateway_options["user_id"] == "explicit-user"
+    assert runner.md_front == "tcp://explicit-md-front"
+    assert runner.broker_id == "explicit-broker"
+    assert runner.user_id == "explicit-user"
 
 
 def test_live_runner_syncs_client_broker_order_id_mapping() -> None:
